Restructure task naming responsibilities
