--- distributed/shuffle/_utils.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from typing import NewType
-
-ShuffleId = NewType("ShuffleId", str)
